refactor: update import paths for Issue interface to use the correct directory
